Implemented the mip offsets.
git-svn-id: https://192.168.3.139:8443/svn/Medical/trunk@5765 956a8c8d-885e-cb46-a0b9-a2f355b4c364

+++ Standalone/VirtualTexture/OriginalTextureInfo.cs
@@ -8,7 +8,7 @@ namespace Medical
 {
     class OriginalTextureInfo
     {
-        public OriginalTextureInfo(String textureUnit, String textureFileName, int mipOffset)
+        public OriginalTextureInfo(String textureUnit, String textureFileName, byte mipOffset)
         {
             this.TextureUnit = textureUnit;
             this.TextureFileName = textureFileName;
@@ -19,6 +19,6 @@ namespace Medical
 
         public String TextureFileName { get; private set; }
 
-        public int MipOffset { get; private set; }
+        public byte MipOffset { get; private set; }
     }
 }
